+++ settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "demo2"
+rootProject.name = "backend"
