+++ src/main/kotlin/com/automatemic/backend/Booking/Booking.kt
@@ -1,21 +1,14 @@
-package com.automatemic.demo2.Booking
+package com.automatemic.backend.Booking
 
-import com.automatemic.demo2.Customer.Customer
-import com.automatemic.demo2.Venue.Venue
-import liquibase.pro.packaged.A
-import lombok.AllArgsConstructor
-import lombok.Data
+import com.automatemic.backend.Customer.Customer
+import com.automatemic.backend.Venue.Venue
 import lombok.NoArgsConstructor
-import java.util.*
-import javax.persistence.Column
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType
 import javax.persistence.Id
 import javax.persistence.JoinColumn
-import javax.persistence.ManyToMany
 import javax.persistence.ManyToOne
-import javax.persistence.OneToMany
 
 @Entity(name = "booking")
 @NoArgsConstructor
