+++ src/main/kotlin/com/automatemic/backend/Booking/BookingController.kt
@@ -1,4 +1,4 @@
-package com.automatemic.demo2.Booking
+package com.automatemic.backend.Booking
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.GetMapping
