+++ src/main/kotlin/com/automatemic/backend/Booking/BookingRepositoryCustomImpl.kt
@@ -1,6 +1,5 @@
-package com.automatemic.demo2.Booking
+package com.automatemic.backend.Booking
 
-import com.automatemic.demo2.Customer.Customer
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Repository
 import javax.persistence.EntityManager
