+++ src/main/kotlin/com/automatemic/backend/Customer/CustomerService.kt
@@ -1,4 +1,4 @@
-package com.automatemic.demo2.Customer
+package com.automatemic.backend.Customer
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
