+++ src/main/kotlin/com/automatemic/backend/Venue/VenueRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.automatemic.backend.Venue
+
+class VenueRepositoryImpl {
+}
