+++ src/main/kotlin/com/automatemic/backend/Venue/VenueService.kt
@@ -1,4 +1,4 @@
-package com.automatemic.demo2.Venue
+package com.automatemic.backend.Venue
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
