--- src/main/kotlin/com/automatemic/demo2/Venue/VenueRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package com.automatemic.demo2.Venue
-
-class VenueRepositoryImpl {
-}
